backend/requirements.txt
# Core Flask
Flask>=3.0.0
Werkzeug>=3.0.0

# Extensions
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.0
python-socketio>=5.11.0

# Supabase
supabase>=2.3.0
postgrest>=0.16.0

# Environment
python-dotenv>=1.0.0

# Server (for production)
gunicorn>=22.0.0

# Validation & Security
email-validator>=2.1.0
phonenumbers>=8.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
Pillow>=10.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2024.1